hmpb/render-backend: Move files out of `next` dir
Also create a `preview` dir for browser preview specific files
