+++ libs/hmpb/render-backend/src/index.ts
@@ -1,3 +1,9 @@
 export * from './all_bubble_ballot_fixtures';
 export * from './ballot_fixtures';
-export * from './next';
+export * from './concatenate_pdfs';
+export * from './mark_ballot';
+export * from './playwright_renderer';
+export * from './render_ballot';
+export * from './renderer';
+export * from './vx_default_ballot_template';
+export * from './types';
