+++ libs/hmpb/render-backend/src/svg_assets.tsx
